@@ -2,6 +2,7 @@ package com.squeaker.games.squeakerapi.controller;
 
 import com.squeaker.games.squeakerapi.model.Game;
 import com.squeaker.games.squeakerapi.service.GamesService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,14 +11,11 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import java.util.List;
 
 @Controller
-@RequestMapping(path = "games")
+@RequestMapping("games")
+@RequiredArgsConstructor
 public class GamesController {
 
-    public GamesController(GamesService service) {
-        this.service = service;
-    }
-
-    private GamesService service;
+    private final GamesService service;
 
     @GetMapping
     public @ResponseBody List<Game> games() {
