@@ -3,10 +3,15 @@ package com.squeaker.games.squeakerapi.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.elasticsearch.annotations.Document;
 
-@AllArgsConstructor
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
+@Document(indexName = "games")
 public class Game {
+    @Id
+    private String id;
     private String name;
 }
