@@ -1,15 +1,19 @@
 package com.squeaker.games.squeakerapi.service;
 
 import com.squeaker.games.squeakerapi.model.Game;
+import com.squeaker.games.squeakerapi.repository.GamesRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class GamesService {
 
+    private final GamesRepository repository;
+
     public List<Game> getGames() {
-        throw new NotImplementedException();
+        return repository.findAll();
     }
 }
